+++ docs/source/Changelog.rst
@@ -35,6 +35,8 @@ Added
 
    -  ``schematic_sheet_name_workaround`` to workaround SHEETNAME on
       page 1 (See #933)
+   -  ``pcb_material_color`` default color for the PCB core material
+   -  ``solder_paste_color`` default color for the solder paste
 
 -  Variants:
 
@@ -57,6 +59,11 @@ Added
 
    -  ``bom`` configuration
    -  ``kiri`` configuration
+
+-  Render 3D:
+
+   -  KiCad 10: support for kicad-cli interface, adds some new options
+      but has various limitations, hopping they will be fixed.
 
 -  Report: